Extract metronome from `OsuModTarget`
